Refactoring and renaming of GameManager into AppManager
