+++ SimonVR/Assets/Scripts/AppManagement/StateMachine/GameOverState.cs
@@ -4,11 +4,11 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace SimonVR.Assets.Scripts.GameManager.StateMachine
+namespace SimonVR.Assets.Scripts.AppManagement.StateMachine
 {
-    public class PlayState : State
+    public class GameOverState : State
     {
-        public PlayState(GameManager gameManager) : base(gameManager)
+        public GameOverState(AppManager gameManager) : base(gameManager)
         { }
 
         public override event EventHandler<State> ChangeStateRequestEvent;
