+++ SimonVR/Assets/Scripts/AppManagement/StateMachine/MenuState.cs
@@ -4,13 +4,13 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace SimonVR.Assets.Scripts.GameManager.StateMachine
+namespace SimonVR.Assets.Scripts.AppManagement.StateMachine
 {
     public class MenuState : State
     {
         public override event EventHandler<State> ChangeStateRequestEvent;
 
-        public MenuState(GameManager gameManager) : base(gameManager)
+        public MenuState(AppManager gameManager) : base(gameManager)
         { }
 
     }
