+++ SimonVR/Assets/Scripts/GameManagement/GameManager.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using UnityEngine;
+
+namespace SimonVR.Assets.Scripts.GameManagement
+{
+    public class GameManager : MonoBehaviour
+    {
+    }
+}
